requirements.txt
aiofiles==0.8.0
aiogoogle==4.2.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Faker==12.0.1
fastapi==0.78.0
fastapi-users==10.0.4
fastapi-users-db-sqlalchemy==4.0.3
flake8==4.0.1
freezegun==1.2.1
frozenlist==1.3.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.22.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
lock==2018.3.25.2110
makefun==1.13.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.2.2
multidict==6.0.2
oauthlib==3.3.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
proto-plus==1.26.1
protobuf==6.31.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.23.4
pytest-freezegun==0.4.2
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
toml==0.10.2
tomli==2.2.1
tonyg-rfc3339==0.1
typing_extensions==4.2.0
uritemplate==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.21.0
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
